refactor: move handlers to frameworks-and-drivers/messaging layer
Move messaging-related handlers from infrastructure/browser to
frameworks-and-drivers/messaging for consistent Clean Architecture
layer structure.

Moved files:
- handlers/background/applyAllRulesHandler.ts
- handlers/content/applyAllRulesHandler.ts
- handlers/content/getElementSelectionHandler.ts
- background/onMessageReceived.ts
- content/onMessageReceived.ts

Also moved related ESLint rules to match the new directory structure.
