+++ host-frontend-root/frontend-src-root/eslint-rules/clean-architecture/main.js
@@ -7,8 +7,8 @@ import domain from '#eslint-rules/clean-architecture/domain.js';
 import frameworksAndDriversMessagingDto from '#eslint-rules/clean-architecture/frameworks-and-drivers/messaging/dto.js';
 import frameworksAndDriversMessagingService from '#eslint-rules/clean-architecture/frameworks-and-drivers/messaging/service.js';
 import frameworksAndDriversUiAtoms from '#eslint-rules/clean-architecture/frameworks-and-drivers/ui/atoms.js';
-import backgroundHandlers from '#eslint-rules/clean-architecture/infrastructure/browser/handlers/background.js';
-import contentHandlers from '#eslint-rules/clean-architecture/infrastructure/browser/handlers/content.js';
+import backgroundHandlers from '#eslint-rules/clean-architecture/frameworks-and-drivers/messaging/handlers/background.js';
+import contentHandlers from '#eslint-rules/clean-architecture/frameworks-and-drivers/messaging/handlers/content.js';
 
 export default [
   ...applicationBusinessRulesDto,
