+++ host-frontend-root/frontend-src-root/src/entrypoints/background.ts
@@ -5,7 +5,7 @@ import {
 import { createRewriteRuleProxyServiceImpl } from 'src/frameworks-and-drivers/messaging/RewriteRuleProxyServiceImpl';
 import { contextMenusOnClicked } from 'src/infrastructure/browser/background/contextMenus/onClicked';
 import { runtimeOnExtensionInstalled } from 'src/infrastructure/browser/background/runtime/onExtensionInstalled';
-import { runtimeOnMessageReceived } from 'src/infrastructure/browser/background/runtime/onMessageReceived';
+import { runtimeOnMessageReceived } from 'src/frameworks-and-drivers/messaging/background/onMessageReceived';
 import { tabsOnUpdated } from 'src/infrastructure/browser/background/tabs/onUpdated';
 
 export default defineBackground({
