+++ host-frontend-root/frontend-src-root/src/frameworks-and-drivers/messaging/background/onMessageReceived.ts
@@ -1,5 +1,5 @@
 import { onBackgroundMessage } from 'src/frameworks-and-drivers/messaging/messaging';
-import { applyAllRulesHandler } from 'src/infrastructure/browser/handlers/background/applyAllRulesHandler';
+import { applyAllRulesHandler } from 'src/frameworks-and-drivers/messaging/handlers/background/applyAllRulesHandler';
 
 /**
  * 呼び出し元: entrypoints/background.ts
