+++ host-frontend-root/frontend-src-root/src/frameworks-and-drivers/messaging/handlers/content/applyAllRulesHandler.ts
